Move TPartitionLocation back to control_plane.h
